Pass ASAHI, CUDA, HIP, HSA prefixed env vars to generated files
Relates-to: #525
